requirements.txt
PyQt5>=5.15
opencv-python>=4.8
face-recognition>=1.3
numpy>=1.24
mysql-connector-python>=8.0
pandas>=2.0
bcrypt>=4.0
setuptools<81

